Split pkg/plugins into per-concern files
registry.go had grown to 979 lines holding four unrelated concerns, and
the two plugin sources were laid out asymmetrically: the directory
source had disk.go to itself while the ConfigMap source sat in
registry.go alongside everything shared. Give each concern its own
file, so the two sources now mirror each other.

  zip.go        archive extraction, and the path-safety primitives that
                exist because entry names and URL segments are untrusted
  configmaps.go the ConfigMap source, mirroring disk.go slot for slot
  manifest.go   manifest schema validation and route-path reservation
  registry.go   the registry itself: entries, caching, merge, Index/File

Tests follow the same split, with shared helpers in helpers_test.go and
the cross-source resolution cases beside the code that resolves them.
Registry.Run also becomes RunConfigMapWatch: an internal Go API change
only, with the HTTP routes, config keys and manifest schema untouched.

Every moved declaration is byte-identical to before, verified by diffing
a name-sorted dump of the package's declarations across the move. The
deliberate exceptions: comments citing a file name the move invalidated,
two cross-source tests renamed to their new file's convention, and
extractZip's tests, which no longer go through parsePluginArchive.
